Support the INTEL compiler
The CPP library, and makefiles, now support using the INTEL tool chain.

A new TOOL_CHAIN can be set to either GNU or INTEL

+++ CPP/Libraries/libUtils/binaryDescribe.cpp
@@ -25,7 +25,7 @@ std::string BinaryDescribe::Describe(const BinaryReader& src, long len) {
             reader+=1;
 
             // Write the binary #
-            unsigned short n = c - 0;
+            int n = c - 0;
             binstr << hex << setw(2) << setfill('0') << n;
             binstr << " ";
